Make changes according to #162#issuecomment-1939463285
